+++ flutter_quill_extensions/lib/embeds/toolbar/camera_button.dart
@@ -19,6 +19,7 @@ class CameraButton extends StatelessWidget {
     this.webVideoPickImpl,
     this.cameraPickSettingSelector,
     this.iconTheme,
+    this.tooltip,
     Key? key,
   }) : super(key: key);
 
@@ -42,6 +43,7 @@ class CameraButton extends StatelessWidget {
   final MediaPickSettingSelector? cameraPickSettingSelector;
 
   final QuillIconTheme? iconTheme;
+  final String? tooltip;
 
   @override
   Widget build(BuildContext context) {
@@ -53,6 +55,7 @@ class CameraButton extends StatelessWidget {
 
     return QuillIconButton(
       icon: Icon(icon, size: iconSize, color: iconColor),
+      tooltip: tooltip,
       highlightElevation: 0,
       hoverElevation: 0,
       size: iconSize * 1.77,